Further refactoring of CDI tests
